fix: date input behaviour in home and room-details pages

+++ database/migrations/2024_09_24_092331_create_rooms_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('rooms', function (Blueprint $table) {
             $table->id();
-            $table->integer('status_id');
+            $table->unsignedBigInteger('status_id');
             
             $table->integer('number');
             $table->text('description');
